# frontend/src/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cortex API",
    "description": "Evidence-first operational intelligence platform for supply chains",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/readiness/batches/{batch_id}": {
      "post": {
        "tags": [
          "readiness"
        ],
        "summary": "Compute Readiness Endpoint",
        "description": "Compute readiness for a batch. Caller is authorized for `workspace_id`.",
        "operationId": "compute_readiness_endpoint_api_v1_readiness_batches__batch_id__post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "pattern": "^[A-Za-z0-9._\\-:]{1,64}$",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/readiness/assessments/{assessment_id}": {
      "get": {
        "tags": [
          "readiness"
        ],
        "summary": "Get Readiness",
        "description": "Get a readiness assessment by ID, scoped to the requested workspace.",
        "operationId": "get_readiness_api_v1_readiness_assessments__assessment_id__get",
        "parameters": [
          {
            "name": "assessment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Assessment Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "pattern": "^[A-Za-z0-9._\\-:]{1,64}$",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReadinessResponse": {
        "properties": {
          "assessment_id": {
            "type": "string",
            "title": "Assessment Id"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "blocking_conflict_count": {
            "type": "integer",
            "title": "Blocking Conflict Count"
          },
          "open_conflict_count": {
            "type": "integer",
            "title": "Open Conflict Count"
          },
          "accepted_claim_count": {
            "type": "integer",
            "title": "Accepted Claim Count"
          },
          "pending_claim_count": {
            "type": "integer",
            "title": "Pending Claim Count"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "assessment_id",
          "workspace_id",
          "batch_id",
          "state",
          "blocking_conflict_count",
          "open_conflict_count",
          "accepted_claim_count",
          "pending_claim_count",
          "assumptions",
          "explanation"
        ],
        "title": "ReadinessResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}